requirements/requirements.txt
Django==1.6.5
argparse==1.2.1
django-debug-toolbar==1.2.1
names==0.3.0
python-dateutil==2.2
six==1.7.3
sqlparse==0.1.11
wsgiref==0.1.2
